Another batch of msvc compile fixes. Added wrapper implementation for opendir/readdir Other functions left stubbed out. It's a shame that crawl's "direct.h" conflicts with <direct.h>. Fixed up use of AppHdr.h in a couple places (it must be included first); changed project to use precompiled headers.
crawl now compiles cleanly but doesn't link.



git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3573 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/religion.cc
@@ -3613,7 +3613,7 @@ void offer_items()
         if (!yesno("Do you wish to part with all of your money? ", true, 'n'))
             return;
 
-        int donation_value = (int) (you.gold/200 * log(you.gold));
+        int donation_value = (int) (you.gold/200 * log((double)you.gold));
 #if DEBUG_DIAGNOSTICS || DEBUG_SACRIFICE || DEBUG_PIETY
         mprf(MSGCH_DIAGNOSTICS, "A donation of $%d amounts to an "
              "increase of piety by %d.", you.gold, donation_value);
